requirements.txt
pillow==9.4.0
pandas==2.0.3
numpy==1.25.2
